######  #50074 #90849565 9 9173783467 153974/2020   IRT-151, BIGGER THAT HSI-55. DEFINITELY.
---
NYSCEF COURT DOCKETS

 260 	EXHIBIT(S)  - 1  (Motion #001)
	
6.24.2020 email - ASHLEY.HUMPHRIES@WILSONELSER.COM   
- DISTRIBUTE THOSE VIDEO.MOV FILES WITHOUT MY CONSENT -- VIOLATION OF PRIVACY AND CIVIL RIGHTS
- AS A REAL ESTATE ATTORNEY IN NEW YORK CIVIL PART 60, UNDER THE HON. TINGLING.

	https://iapps.courts.state.ny.us/nyscef/ViewDocument?docIndex=ANDSuLFWlbjttGxgjUXGAw==
  
 261 	EXHIBIT(S)  - 2  (Motion #001)

6.25.2020 email-ASHLEY.HUMPHRIES@WILSONELSER.COM   - DISTRIBUTE THOSE VIDEO.MOV FILES WITHOUT MY CONSENT 
-- VIOLATION OF PRIVACY AND CIVIL RIGHTS
	https://iapps.courts.state.ny.us/nyscef/ViewDocument?docIndex=ph76T77EGjUtofiKDKCOoA==
  
262 	EXHIBIT(S)  - 3  (Motion #001)

6.27.2020 email-ASHLEY.HUMPHRIES@WILSONELSER.COM   - DISTRIBUTE THOSE VIDEO.MOV FILES WITHOUT MY CONSENT 
-- VIOLATION OF PRIVACY AND CIVIL RIGHTS
	https://iapps.courts.state.ny.us/nyscef/ViewDocument?docIndex=x9RFyUAiwcmGbb6y8R306w==
  
 263 	EXHIBIT(S)  - 4  (Motion #001)

6.27.2020 email-ASHLEY.HUMPHRIES@WILSONELSER.COM   - DISTRIBUTE THOSE VIDEO.MOV FILES WITHOUT MY CONSENT 
-- VIOLATION OF PRIVACY AND CIVIL RIGHTS
	https://iapps.courts.state.ny.us/nyscef/ViewDocument?docIndex=hxv_PLUS_e1_PLUS_miua_PLUS_MYiCxhv_PLUS_XQ==
 264 	EXHIBIT(S)  - 5  (Motion #001)

6.27.2020 email-ASHLEY.HUMPHRIES@WILSONELSER.COM   - DISTRIBUTE THOSE VIDEO.MOV FILES WITHOUT MY CONSENT 
-- VIOLATION OF PRIVACY AND CIVIL RIGHTS
	https://iapps.courts.state.ny.us/nyscef/ViewDocument?docIndex=382TNrNKjw28ce0sCmTN9g==
 265 	EXHIBIT(S)  - 6  (Motion #001)

6.28.2020 email-ASHLEY.HUMPHRIES@WILSONELSER.COM   - DISTRIBUTE THOSE VIDEO.MOV FILES WITHOUT MY CONSENT 
-- VIOLATION OF PRIVACY AND CIVIL RIGHTS
	https://iapps.courts.state.ny.us/nyscef/ViewDocument?docIndex=UQzt9l/yhPj2jXV6Xk9JEA==
 266 	EXHIBIT(S)  - 7  (Motion #001)

6.29.2020 email-ASHLEY.HUMPHRIES@WILSONELSER.COM   - DISTRIBUTE THOSE VIDEO.MOV FILES WITHOUT MY CONSENT 
-- VIOLATION OF PRIVACY AND CIVIL RIGHTS
	https://iapps.courts.state.ny.us/nyscef/ViewDocument?docIndex=ZWQqfKHMbUj/LLDfoydj1g==
 267 	EXHIBIT(S)  - 8  (Motion #001)

6.30.2020 email
	https://iapps.courts.state.ny.us/nyscef/ViewDocument?docIndex=bJO1Rhql6hYHUqwyBXgpGQ==
![image](https://user-images.githubusercontent.com/110014688/180958129-3acb7a14-6c45-49ef-b4fd-9ef88e105668.png)



*** THIS INFORMATION WILL BE MAINTAINED, AND IT IS UNDERSTOOD THAT ANY INFORMATION IS MAINTAINED IN A RESTRISTED SYSTEM ARCHIVE, THE INFORMATION SUPPLIED IN PRACTICE, AS AN EXPERT WAS VOLUNTEERED UNDER #50074 90849565 TO BRING AN EXPEDITED PROCEDURE AND A CLOSURE TO THE CONTINUING FINANCIAL CRIMES, AND OTHER CRIMES AS DISCUSSED WITH THE REPRESENTATIVES FROM SEVERAL FEDERAL AGENCIES AND PRIOR TO INITIATING THIS FACTION AND UPON APPROVAL, HAVE NOT BREACHED ANY 


The collection of information, logging of use of activity is authorized under 28 USC Section 534, 42 USC Section 3301, and under the COMPUTER SECURITY ACT of 1987, the general record keeping provision of the Administrative Procedures Act ( 5 USC Section 301), and the Federal Information Security Management Act of 2002.

The files and information (e.g. phone numbers, email address, and agency affilian) permits the FBI to verify your identity and confirms yourself as a qualified user, at their discretion. Any information collected will be stored by the respective agencies where applicable, including Personally Identifiable Information ( PII ) with other agencies that have been included in this matter.

By visiting this information, you understand that you also consent to the collection of information which is collected by all collaborators, notwithstanding other Federal Bureau of Investigations ( FBI ) criminal justice partners that are deemed as approved in their capacity, and to provide a real time environment to communicate this information by and between those experts and by and between the faction of common interests which are confidential, and under the Authority of the FBI are governed in this matter.

As permitted by the Privacy Act of 1974, users that are provided with a username and password, or otherwise access this information have accepted that without consent, may be distributed as required by federal or state law, as permitted by the Privacy Act of 1974.

Below is a listing of systems of records maintained by DOJ. A system of records is defined by the Privacy Act of 1974 as a group of any records under the control of any agency from which information is retrieved by the name of the individual or by some identifying number, symbol, or other identifying particular assigned to the individual.  Rules exempting systems of records from certain Privacy Act requirements are in 28 CFR Part 16, Subpart E, and are listed with the corresponding system of records, below, for convenience:

[The following agencies, notwithstanding others may or may not be monitoring this information.](https://www.justice.gov/opcl/privacy-act-1974#:~:text=The%20Privacy%20Act%20of%201974,of%20records%20by%20federal%20agencies.)



[Exemptions, where applicable](https://www.ecfr.gov/current/title-28/chapter-I/part-16#sp28.1.16.e) prior to review of any information herein.
![image](https://user-images.githubusercontent.com/110014688/180948172-b911ea8a-d983-42dc-908a-5b2a610929cf.png)